pyproject.toml
[project]
name = "workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.7.11",
    "arxiv>=2.2.0",
    "beautifulsoup4>=4.13.4",
    "ddgs>=9.5.2",
    "duckduckgo-search>=8.1.1",
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.30.0",
    "google-generativeai>=0.8.5",
    "googlesearch-python>=1.3.0",
    "openai>=1.99.9",
    "pycountry>=24.6.1",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.43",
    "yfinance>=0.2.65",
]
